+++ examples/simple_classifier_cifar10/start_without_hpo.py
@@ -3,7 +3,7 @@
 
 from syne_tune.backend.sagemaker_backend.sagemaker_utils import get_execution_role
 
-from renate.run import run_training_job
+from renate.training import run_training_job
 
 config_space = {
     "optimizer": "SGD",
