+++ examples/train_mlp_locally/start_training_with_er_without_hpo.py
@@ -1,4 +1,4 @@
-from renate.run import run_training_job
+from renate.training import run_training_job
 
 configuration = {
     "optimizer": "SGD",
