+++ src/renate/training/__init__.py
@@ -1,5 +1,5 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
-from renate.run.training_job import run_training_job
+from renate.training.training_job import run_training_job
 
 __all__ = ["run_training_job"]
